hw 2

+++ src/homework/c_decisions/main.py
@@ -3,8 +3,8 @@ import decisions
 
 def main():
     #prompts user for options
-    options = int(input ('Enter an option: '))
-    total = int(input ('Enter a total: '))
+    options = float(input ('Enter an option: '))
+    total = float(input ('Enter a total: '))
 
     #creates a result variable from options ratio function
     result = decisions.get_options_ratio(options, total)
